backend/requirements.txt
# ==================== PYTHON ====================

# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-multipart>=0.0.20
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=24.1.0

# Rate Limiting & Security
slowapi>=0.1.9  # 限流
python-hcl2>=0.8.0

# Vector & Search
qdrant-client>=1.12.0
bm25s>=0.2.2

# LLM
openai>=1.55.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
anthropic>=0.40.0
google-generativeai>=0.8.0

# Knowledge Graph
networkx>=3.4
neo4j>=5.26.0  # 可选

# Document Parsing
python-docx>=1.1.2
markdown>=3.7
beautifulsoup4>=4.12.3
python-magic>=0.4.27

# Utils
python-dotenv>=1.0.1
loguru>=0.7.2
tiktoken>=0.8.0
httpx>=0.28.0
orjson>=3.10.0  # 快速 JSON
cachetools>=5.5.0
python-dateutil>=2.9.0

# Redis (可选)
redis>=5.2.0
celery>=5.5.0

# Monitoring & Error Tracking
sentry-sdk>=2.0.0  # 错误追踪
prometheus-client>=0.20.0  # 指标暴露

# Backup & SSL
cryptography>=42.0.0  # SSL 证书

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
httpx>=0.28.0
